backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
contextlib2==21.6.0
cryptography==46.0.5
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
django-ninja==1.6.1
django-ninja-extra==0.31.2
django-ninja-jwt==5.4.4
gunicorn==25.1.0
idna==3.11
injector==0.24.0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
